:hammer: Moved `getStyle` functions to separate file.
